+++ engine/source/webmlib/vidplay.h
@@ -30,7 +30,7 @@ webm_context *webm_start_playback_ex(
 void webm_get_video_info(webm_context *ctx, yuv_video_mode *dims);
 uint64_t webm_get_duration(webm_context *ctx);
 void webm_set_audio_paused(webm_context *ctx, int paused);
-void webm_set_audio_speed(webm_context *ctx, unsigned int speed);
+void webm_set_audio_speed(webm_context *ctx, double speed);
 void webm_close(webm_context *ctx);
 int webm_try_get_next_frame(webm_context *ctx, yuv_frame **frame);
